+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 ## Bug Fixes
 
+- [#91](https://github.com/pybop-team/PyBOP/issues/91) - Adds a check on the number of parameters for CMAES and makes XNES the default optimiser.
+
 # [v24.3](https://github.com/pybop-team/PyBOP/tree/v24.3) - 2024-03-25
 
 ## Features